requirements.txt
amqp==5.0.1
aniso8601==8.0.0
billiard==3.6.3.0
celery==5.0.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
coverage==5.3
Flask==1.1.2
Flask-Caching==1.9.0
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
mongoengine==0.20.0
nose==1.3.7
prompt-toolkit==3.0.8
pycryptodome==3.9.8
pymongo==3.11.0
python-memcached==1.59
pytz==2020.1
redis==3.5.3
rednose==1.3.0
six==1.15.0
termstyle==0.1.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.3.0
